add CI and cost to the confidence set (#14)
add 95% CI and cost associated for each row of the confidence set. 
please note that I refactored a lot of code from
`calculate_recommended_intervention.R`, so that function is a lot
cleaner now. I didn't ask for your review for the refactoring PRs as the
code pretty much stayed the same.

To test, I ran all of the Rmd files (except `test_pulesa_cost_related`
which i haven't got to it yet), and the results look correct to me.