shrinking among other improvements (#26)
1. add IQR of the cost of the final recommended package 
2. add confidence interval for the estimated outcome goal, using the
final recommended intervention,
3. add P value of the overall two sample test, overall intervention
test, if user specify ‘group’ column. only supports binary outcomes for
now.
4. add helper function to implement the shrinking method   
5. add the logic for calling the shrinking method: When outcome goal
cannot be reached, and the estimated maximum achievable outcome goal is
less than 1/4 of the distance from the estimated outcome without
intervention to the per-specified outcome goal, then we shrink. We
inform the user about this 1/4 value, if they want to change it they can
do so.
6. for the cost function visualization, add visual warnings when total
cost functions are not non-decreasing anymore.
7. based on discussion with Jingyu, different starting values are needed
for non linear solution, need to compare the results based on different
starting values, and select the best one.